Backend/requirements.txt
aiofiles==23.2.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
cryptography==46.0.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
google-ai-generativelanguage==0.6.4
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-generativeai==0.5.4
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.4
orjson==3.11.8
pandas==2.2.2
proto-plus==1.27.2
protobuf==4.25.9
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
rich==14.3.3
rich-toolkit==0.19.7
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.30
starlette==0.37.2
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
